requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
defusedxml==0.5.0
flake8==3.4.1
Flask==0.12.2
idna==2.6
image==1.5.16
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
numpy==1.13.1
olefile==0.44
pandas==0.20.3
Pillow==4.3.0
pkg-resources==0.0.0
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
wordcloud==1.3.1
